requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.50.0
pillow>=9.0.0
numpy>=1.21.0
requests>=2.25.0
accelerate>=0.20.0
